Renamed video_process to video_dump
