+++ src/video_dump.c
@@ -1,4 +1,4 @@
-#include "video_process.h"
+#include "video_dump.h"
 #include "video.h"
 #include <sys/stat.h>
 #include <fcntl.h>
